Split out the _xxinterpchannels module.
