+++ configure
@@ -25615,6 +25615,7 @@ case $ac_sys_system in #(
     py_cv_module__scproxy=n/a
     py_cv_module__tkinter=n/a
     py_cv_module__xxsubinterpreters=n/a
+    py_cv_module__xxinterpchannels=n/a
     py_cv_module_grp=n/a
     py_cv_module_nis=n/a
     py_cv_module_ossaudiodev=n/a
@@ -26053,6 +26054,26 @@ fi
 
   as_fn_append MODULE_BLOCK "MODULE__XXSUBINTERPRETERS_STATE=$py_cv_module__xxsubinterpreters$as_nl"
   if test "x$py_cv_module__xxsubinterpreters" = xyes; then :
+
+
+
+
+fi
+
+
+        if test "$py_cv_module__xxinterpchannels" != "n/a"; then :
+  py_cv_module__xxinterpchannels=yes
+fi
+   if test "$py_cv_module__xxinterpchannels" = yes; then
+  MODULE__XXINTERPCHANNELS_TRUE=
+  MODULE__XXINTERPCHANNELS_FALSE='#'
+else
+  MODULE__XXINTERPCHANNELS_TRUE='#'
+  MODULE__XXINTERPCHANNELS_FALSE=
+fi
+
+  as_fn_append MODULE_BLOCK "MODULE__XXINTERPCHANNELS_STATE=$py_cv_module__xxinterpchannels$as_nl"
+  if test "x$py_cv_module__xxinterpchannels" = xyes; then :
